# src/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.2"
  title: Biomodels API Services
# during dev, should point to your local machine
host: localhost:4000
...
basePath: /biomodelos
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /species:
    # binds a127 app logic to a route
    x-swagger-router-controller: biomodels
    get:
      description: "TO DO"
      operationId: species
      responses:
        "200":
          description: Success
          schema:
            type: array
            items:
              type: object
              properties:
                taxID:
                  type: integer
                species:
                  type: string
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string